add assignor package and move the classes into the package
