+++ common/src/main/java/org/astraea/common/consumer/assignor/Assignment.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.astraea.common.consumer.assignor;
+
+import java.util.List;
+import java.util.Map;
+import org.astraea.common.admin.TopicPartition;
+
+public final class Assignment {
+  private List<TopicPartition> partitions;
+  private Map<String, String> userData;
+
+  public Assignment(List<TopicPartition> partitions, Map<String, String> userData) {
+    this.partitions = partitions;
+    this.userData = userData;
+  }
+
+  public Assignment(List<TopicPartition> partitions) {
+    this(partitions, null);
+  }
+
+  public List<TopicPartition> partitions() {
+    return partitions;
+  }
+
+  public Map<String, String> userData() {
+    return userData;
+  }
+
+  @Override
+  public String toString() {
+    return "Assignment("
+        + "partitions="
+        + partitions
+        + (userData == null ? "" : ", user data= " + userData)
+        + ')';
+  }
+}
