+++ common/src/main/java/org/astraea/common/consumer/assignor/ConsumerPartitionAssignor.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.astraea.common.consumer.assignor;
+
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import org.apache.kafka.common.Cluster;
+import org.astraea.common.admin.ClusterInfo;
+import org.astraea.common.admin.ReplicaInfo;
+import org.astraea.common.admin.TopicPartition;
+
+public interface ConsumerPartitionAssignor
+    extends org.apache.kafka.clients.consumer.ConsumerPartitionAssignor {
+
+  /**
+   * Perform the group assignment given the member subscriptions and current cluster metadata.
+   *
+   * @param subscriptions Map from the member id to their respective topic subscription.
+   * @param metadata Current topic/broker metadata known by consumer.
+   * @return Map from each member to the list of partitions assigned to them.
+   */
+  Map<String, List<TopicPartition>> assign(
+      Map<String, org.astraea.common.consumer.assignor.Subscription> subscriptions,
+      ClusterInfo<ReplicaInfo> metadata);
+
+  @Override
+  default String name() {
+    return "astraea";
+  }
+
+  @Override
+  default GroupAssignment assign(Cluster metadata, GroupSubscription groupSubscription) {
+
+    // convert Kafka's data structure to ours
+    var clusterInfo = ClusterInfo.of(metadata);
+    var subscriptionsPerMember =
+        org.astraea.common.consumer.assignor.GroupSubscription.from(groupSubscription)
+            .groupSubscription();
+
+    return new GroupAssignment(
+        assign(subscriptionsPerMember, clusterInfo).entrySet().stream()
+            .collect(
+                Collectors.toMap(
+                    Map.Entry::getKey,
+                    e ->
+                        new Assignment(
+                            e.getValue().stream()
+                                .map(TopicPartition::to)
+                                .collect(Collectors.toUnmodifiableList())))));
+  }
+}
