+++ common/src/main/java/org/astraea/common/consumer/assignor/GroupAssignment.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.astraea.common.consumer.assignor;
+
+import java.util.Map;
+
+public final class GroupAssignment {
+  private final Map<String, Assignment> assignments;
+
+  public GroupAssignment(Map<String, Assignment> assignments) {
+    this.assignments = assignments;
+  }
+
+  public Map<String, Assignment> groupAssignment() {
+    return assignments;
+  }
+
+  @Override
+  public String toString() {
+    return "GroupAssignment(" + "assignments=" + assignments + ")";
+  }
+}
