+++ common/src/test/java/org/astraea/common/consumer/ConsumerTest.java
@@ -593,7 +593,7 @@ public class ConsumerTest extends RequireBrokerCluster {
             .bootstrapServers(bootstrapServers())
             .config(
                 ConsumerConfigs.PARTITION_ASSIGNMENT_STRATEGY_CONFIG,
-                org.astraea.common.consumer.RandomAssignor.class.getName())
+                org.astraea.common.consumer.assignor.RandomAssignor.class.getName())
             .seek(DISTANCE_FROM_BEGINNING, (long) 0)
             .build()) {
       var records = consumer.poll(30, Duration.ofSeconds(5));
@@ -660,7 +660,7 @@ public class ConsumerTest extends RequireBrokerCluster {
                                       .config(ConsumerConfigs.GROUP_ID_CONFIG, groupId)
                                       .config(
                                           ConsumerConfigs.PARTITION_ASSIGNMENT_STRATEGY_CONFIG,
-                                          org.astraea.common.consumer.RandomAssignor.class
+                                          org.astraea.common.consumer.assignor.RandomAssignor.class
                                               .getName())
                                       .bootstrapServers(bootstrapServers())
                                       .seek(SEEK_TO, 0)
